Move users request models to local 'api_models' package
